requirements.txt
# UI
streamlit

# API Key
python-dotenv

# LLM
openai
langchain
langchain-experimental
